Removes camera functionality from vision and now it has its own subsystem
